Use q launch defaults for dcla for com.android.resolv.
DCLA (Dynamic Common Library Apex) is a mechanism used by Mainline, which places shared c++ libraries into a separate apex in order to de-duplicate them, with the goal of reducing the overall size of a train.

This change results in two packages being built:
If the APEX_BUILD_FOR_PRE_S_DEVICES=1 is set, the apex will have min_sdk set to 29 as before, and this will deliver to Android 11.
If the flag is not set, the second package will have min_sdk set to 31, and will deliver to Android 12 and beyond.

If developers intend to build an apex for use in Q as before, they must set the flag APEX_BUILD_FOR_PRE_S_DEVICES=1, which will set the min_sdk to 29 as before. If not set, the new default behavior will set the min_sdk to 31.

This new variant will be post-processed and trimmed by Mainline infrastructure, and the resulting trimmed module can only be installed on S+ devices, requiring that we set the min_sdk to 31.

Ignore-AOSP-First: Presubmit in AOSP broken due to b/247469919, will cherry-pick to AOSP shortly

Test: APEX_BUILD_FOR_PRE_S_DEVICES=1 m com.android.resolv & verify
      sdk_version = 29
Test: m com.android.resolv & verify sdk_version = 31
Change-Id: I9a74fabe5e787d1c977af7fd7e78bd52e19ebbcc
(cherry picked from commit 5d28d74fdc17288f86761d9795e02fcf9201424c)
Merged-In: I9a74fabe5e787d1c977af7fd7e78bd52e19ebbcc