Correctly support AI_ADDRCONFIG for mDNS queries via getaddrinfo
Current code assumes that having a default gateway is a required to send
out AAAA queries. This is true when using conventional unicast DNS
server, but not when using mDNS. In an mDNS scenario, we can have an
interface with a link-local IPv6 address assigned, but no default
gateway (e.g., IPv6 over Bluetooth). So, under these circumstances, we
are not sending out an AAAA query, even though we could have local IPv6
connectivity.

Bug: b:374756136
Change-Id: I2a4c945cac7b25a6828d232615ace14b61b75304