Refactor dns_getaddrinfo to support more than two queries
Current code assumes that at most two queries are created for each
call to dns_getaddrinfo. This is currently true, but will no longer be
the case as we introduce proper mDNS support: mDNS requires sending
queries on multiple interfaces, breaking the original assumption.

This should effectively be a no-op from the perspective of getaddrinfo.

Bug: b:374756136
Change-Id: I3ba2713d262d18bcc93a7bda6dda7ec1335818da