Always add "lo" as an interface to network TEST_NETID
mDNS DnsResolver tests currently forward all multicast traffic to the
loopback interface (see SetMdnsRoute), but that interface never is
configured to be part of TEST_NETID.
Follow-up changes to mDNS will require a network to have an associated
interface, otherwise no queries will be sent out.
To prepare for that, add the loopback interface to the network used for
testing. Ideally, we could get rid of this and redirect everything
towards a TUN. But setting that up is a lot of work.

Bug: b:374756136
Change-Id: I611b1acc908a3a6c89c0161d7cccb5bdba7218b9