Disable caching for mDNS
Follow-up change will start sending mDNS queries over all interfaces
attached onto a network. Additionally, they could be sent to both
FF02::FB and 224.0.0.251, if the target network support both IPv4 and
IPv6.
In both of these scenarios, we don't want:
1. A query for interface X to wait for a query on interface Y to
   complete
2. A query for interface X to return a cached result from a query to
   interface Y

Ideally, when doing mDNS we would start keying the cache by netid and
interface index, but that is not trivial. As such, to get to a working
prototype sooner, this CL disables caching for mDNS.

Bug: b:374756136, b:394031336
Change-Id: I6d601439a6d9e3432862e5d5a9f52c51b7ec5cda