Send mDNS queries onto every interface associated to target network
The target interface for an mDNS query is represented by
ResState::target_interface_index_for_mdns. It is stored there to
simplify plumbing, as this is already available everywhere (and it
follows the current style of storing sockets/addresses there).

Additionally, if the reply contains link-local IPv6 addresses, record the
interface we used as sin6_scope_id.

Bug: b:374756136
Change-Id: I1b0a53838a84387f371ed73c43dadc4870599b74