getaddrinfo: always populate resources for network
Currently, we call resolv_populate_res_for_net in two scenarios:
1. When the hostname we're trying to resolve doesn't define a domain. In
   this scenario we will try to lookup that hostname within all the
   search domain configured for the target network.
2. While sending out a query, within res_nsend, after a cache miss (not
   due to RESOLV_CACHE_UNSUPPORTED).

The only scenario in which it doesn't end up being called is: a lookup
for a hostname that includes a domain, whose result we have in cache.
This codepath is already grabbing this mutex, grabbing it an additional
time for a cheap operation shuold be okay.

Bug: b:374756136
Change-Id: I847d4106b0f793946a877e7c2de48779a6742c98